mv FT, KV tests to right place
